Replace filter flag with match
The filter flag is not really that useful - using the filter step
in pipeline provides the same functionality without too much overhead.

The match flag takes regular expression to use to filter key/bucket
names. Provides functionality suggested in #6 but as glob syntax is
not compatible with Go regexp it has to be a separate flag.